refactor: 기존 핸들러, 픽스쳐, 타입을 새로운 파일로 덮어씌움
