--- frontend/src/mocks/fixtures/events_new.ts
@@ -1,87 +1,0 @@
-import { MeetingEvent } from 'types/eventType_new';
-
-const events: (MeetingEvent & { meetingId: number })[] = [
-  {
-    id: 0,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-02-08',
-  },
-  {
-    id: 1,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-03-08',
-  },
-  {
-    id: 2,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-04-08',
-  },
-  {
-    id: 3,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-05-08',
-  },
-  {
-    id: 4,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-06-08',
-  },
-  {
-    id: 5,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-07-08',
-  },
-  {
-    id: 6,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-08-08',
-  },
-  {
-    id: 7,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-09-08',
-  },
-  {
-    id: 8,
-    meetingId: 1,
-    attendanceOpenTime: '09:30',
-    attendanceClosedTime: '10:05',
-    meetingStartTime: '10:00',
-    meetingEndTime: '18:00',
-    date: '2022-10-08',
-  },
-];
-
-export default events;
