+++ frontend/src/mocks/handlers/meetingHandler.ts
@@ -1,15 +1,13 @@
 import { DefaultBodyType, rest } from 'msw';
-import meetings from 'mocks/fixtures/meeting';
+import meetings from 'mocks/fixtures/meetings';
 import users from 'mocks/fixtures/users';
-import {
-  AttendanceStatus,
-  UserAttendanceCheckRequestBody,
-  UserCoffeeStatsResponseBody,
-} from 'types/userType';
+import { AttendanceStatus, UserCoffeeStatsResponseBody } from 'types/userType';
 import {
   MeetingListResponseBody,
   MeetingCreateRequestBody,
   MeetingResponseBody,
+  MeetingUpdateRequestBody,
+  MeetingMasterAssignRequestBody,
 } from 'types/meetingType';
 import { DELAY } from 'mocks/configs';
 import { extractIdFromHeader } from 'mocks/utils';
@@ -44,15 +42,22 @@ export default [
     );
 
     const responseBody: MeetingListResponseBody = {
-      meetings: myMeetings.map(({ attendanceCount, userIds, ...meeting }) => ({
-        ...meeting,
-        isMaster: false,
-        isCoffeeTime: true,
-        tardyCount: 3,
-        hasUpcomingEvent: false,
-        entranceTime: '10:00',
-        closingTime: '10:05',
-      })),
+      meetings: myMeetings.map(
+        ({ attendanceEventCount, userIds, ...meeting }) => ({
+          ...meeting,
+          isLoginUserMaster: false,
+          isCoffeeTime: true,
+          tardyCount: 3,
+          upcomingEvent: {
+            id: 1,
+            attendanceOpenTime: '09:30',
+            attendanceClosedTime: '10:05',
+            meetingStartTime: '10:00',
+            meetingEndTime: '18:00',
+            date: '2022-08-08',
+          },
+        })
+      ),
     };
 
     return res(ctx.status(200), ctx.json(responseBody), ctx.delay(DELAY));
@@ -102,13 +107,10 @@ export default [
 
       const { userIds, ...joinedMeeting } = {
         ...meeting,
-        isMaster: true,
+        isLoginUserMaster: true,
         isCoffeeTime,
-        hasUpcomingEvent: true,
         isActive: true,
-        entranceTime: '10:00',
-        leaveTime: '18:00',
-        users: joinedUsers,
+        users: joinedUsers.map((user) => ({ ...user, isMaster: false })),
       };
 
       const responseBody: MeetingResponseBody = joinedMeeting;
@@ -117,50 +119,12 @@ export default [
     }
   ),
 
-  rest.put<UserAttendanceCheckRequestBody, MeetingPathParams>(
-    `${process.env.API_SERVER_HOST}/meetings/:meetingId/users/:userId`,
-    (req, res, ctx) => {
-      const token = extractIdFromHeader(req);
-
-      if (!token.isValidToken) {
-        return res(
-          ctx.status(401),
-          ctx.json({ message: '유효하지 않은 토큰입니다.' })
-        );
-      }
-
-      const user = users.find(({ id }) => id === token.id);
-
-      if (!user) {
-        return res(
-          ctx.status(404),
-          ctx.json({ message: '유저가 존재하지 않습니다.' })
-        );
-      }
-
-      const { meetingId, userId } = req.params;
-      const targetMeeting = meetings.find(
-        (meeting) => meeting.id === Number(meetingId)
-      );
-      if (!targetMeeting) {
-        return res(ctx.status(404), ctx.delay(DELAY));
-      }
-
-      const targetUser = users.find((user) => user.id === Number(userId));
-      if (!targetUser) {
-        return res(ctx.status(404), ctx.delay(DELAY));
-      }
-
-      return res(ctx.status(204), ctx.delay(DELAY));
-    }
-  ),
-
   rest.post<MeetingCreateRequestBody>(
     `${process.env.API_SERVER_HOST}/meetings`,
     (req, res, ctx) => {
       const token = extractIdFromHeader(req);
 
-      if (!token.isValidToken) {
+      if (!token.isValidToken || !token.id) {
         return res(
           ctx.status(401),
           ctx.json({ message: '유효하지 않은 토큰입니다.' })
@@ -177,20 +141,104 @@ export default [
       }
 
       const meeting = req.body;
-      const id = meetings.length;
+      const meetingId = meetings.length;
 
       meetings.push({
         ...meeting,
-        id,
+        id: meetingId,
         isActive: false,
-        attendanceCount: 0,
+        attendanceEventCount: 0,
+        masterId: token.id,
       });
 
       return res(
         ctx.status(201),
-        ctx.set('Location', `/meetings/${id}`),
+        ctx.set('Location', `/meetings/${meetingId}`),
         ctx.delay(DELAY)
       );
+    }
+  ),
+
+  rest.put<MeetingUpdateRequestBody, Pick<MeetingPathParams, 'meetingId'>>(
+    `${process.env.API_SERVER_HOST}/meetings/:meetingId`,
+    (req, res, ctx) => {
+      const token = extractIdFromHeader(req);
+      const meetingId = Number(req.params.meetingId);
+      const { name } = req.body;
+
+      if (!token.isValidToken) {
+        return res(
+          ctx.status(401),
+          ctx.json({ message: '유효하지 않은 토큰입니다.' })
+        );
+      }
+
+      const user = users.find(({ id }) => id === token.id);
+
+      if (!user) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '유저가 존재하지 않습니다.' })
+        );
+      }
+
+      const meeting = meetings.find(({ id }) => id === meetingId);
+
+      if (!meeting) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '모임이 존재하지 않습니다.' })
+        );
+      }
+
+      if (meeting.masterId !== token.id) {
+        return res(ctx.status(403));
+      }
+
+      meeting.name = name;
+
+      return res(ctx.status(204), ctx.delay(DELAY));
+    }
+  ),
+
+  rest.delete<DefaultBodyType, Pick<MeetingPathParams, 'meetingId'>>(
+    `${process.env.API_SERVER_HOST}/meetings/:meetingId`,
+    (req, res, ctx) => {
+      const token = extractIdFromHeader(req);
+      const meetingId = Number(req.params.meetingId);
+
+      if (!token.isValidToken) {
+        return res(
+          ctx.status(401),
+          ctx.json({ message: '유효하지 않은 토큰입니다.' })
+        );
+      }
+
+      const user = users.find(({ id }) => id === token.id);
+
+      if (!user) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '유저가 존재하지 않습니다.' })
+        );
+      }
+
+      const meetingIndex = meetings.findIndex(({ id }) => id === meetingId);
+
+      if (meetingIndex === -1) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '모임이 존재하지 않습니다.' })
+        );
+      }
+
+      if (meetings[meetingIndex].masterId !== token.id) {
+        return res(ctx.status(403));
+      }
+
+      meetings.splice(meetingIndex, 1);
+
+      return res(ctx.status(204), ctx.delay(DELAY));
     }
   ),
 
@@ -245,6 +293,117 @@ export default [
         return res(
           ctx.status(401),
           ctx.json({ message: '유효하지 않은 토큰입니다.' })
+        );
+      }
+
+      return res(ctx.status(204), ctx.delay(DELAY));
+    }
+  ),
+
+  rest.delete<DefaultBodyType, Pick<MeetingPathParams, 'meetingId'>>(
+    `${process.env.API_SERVER_HOST}/meetings/:meetingId/me`,
+    (req, res, ctx) => {
+      const token = extractIdFromHeader(req);
+      const meetingId = Number(req.params.meetingId);
+
+      if (!token.isValidToken) {
+        return res(
+          ctx.status(401),
+          ctx.json({ message: '유효하지 않은 토큰입니다.' })
+        );
+      }
+
+      const meeting = meetings.find(({ id }) => id === meetingId);
+
+      if (!meeting) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '모임이 존재하지 않습니다.' })
+        );
+      }
+
+      if (!meeting.userIds.find((id) => id === token.id)) {
+        return res(
+          ctx.status(404),
+          ctx.json({
+            message: '해당 미팅에 유저가 존재하지 않습니다.',
+          })
+        );
+      }
+
+      if (meeting.masterId === token.id) {
+        return res(
+          ctx.status(403),
+          ctx.json({
+            message: '마스터는 모임을 나갈 수 없습니다.',
+          })
+        );
+      }
+
+      return res(ctx.status(204), ctx.delay(DELAY));
+    }
+  ),
+
+  rest.put<
+    MeetingMasterAssignRequestBody,
+    Pick<MeetingPathParams, 'meetingId'>
+  >(
+    `${process.env.API_SERVER_HOST}/meetings/:meetingId/me`,
+    (req, res, ctx) => {
+      const token = extractIdFromHeader(req);
+      const meetingId = Number(req.params.meetingId);
+      const { userId } = req.body;
+
+      if (!token.isValidToken) {
+        return res(
+          ctx.status(401),
+          ctx.json({ message: '유효하지 않은 토큰입니다.' })
+        );
+      }
+
+      const user = users.find(({ id }) => id === token.id);
+      const targetUser = users.find(({ id }) => id === userId);
+
+      if (!user || !targetUser) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '유저가 존재하지 않습니다.' })
+        );
+      }
+
+      const meeting = meetings.find(({ id }) => id === meetingId);
+
+      if (!meeting) {
+        return res(
+          ctx.status(404),
+          ctx.json({ message: '모임이 존재하지 않습니다.' })
+        );
+      }
+
+      if (meeting.masterId !== token.id) {
+        return res(
+          ctx.status(403),
+          ctx.json({
+            message: '해당 모임의 마스터만 마스터 권한을 양도할 수 있습니다.',
+          })
+        );
+      }
+
+      if (!meeting.userIds.find((id) => id === userId)) {
+        return res(
+          ctx.status(404),
+          ctx.json({
+            message: '해당 미팅에 유저가 존재하지 않습니다.',
+          })
+        );
+      }
+
+      if (meeting.masterId === userId) {
+        return res(
+          ctx.status(403),
+          ctx.json({
+            message: '마스터에게는 마스터 권한을 양도할 수 있습니다.',
+          })
         );
       }
