+++ frontend/src/types/eventType.ts
@@ -1,5 +1,12 @@
 export type MeetingEvent = {
+  id: number;
   date: string;
-  entranceTime: string;
-  leaveTime: string;
+  attendanceOpenTime: string;
+  attendanceClosedTime: string;
+  meetingStartTime: string;
+  meetingEndTime: string;
 };
+
+export type EventResposeBody = MeetingEvent;
+
+export type EventListResponseBody = MeetingEvent[];
