--- frontend/src/types/eventType_new.ts
@@ -1,12 +1,0 @@
-export type MeetingEvent = {
-  id: number;
-  date: string;
-  attendanceOpenTime: string;
-  attendanceClosedTime: string;
-  meetingStartTime: string;
-  meetingEndTime: string;
-};
-
-export type EventResposeBody = MeetingEvent;
-
-export type EventListResponseBody = MeetingEvent[];
